Add background video option (mp4 and webm)
Stop tabs from changing URL (adding hashtag) which prevents the builder from loading if the page is refreshed
Fix bug where all inline settings are sharing an object reference with their parents when first added to a page. This meant that if you add a fresh section / row and change anything in the design tab before saving / reloading, it would change settings for the section, row, and module at the same time.